+++ packages/@storefront/sayt/test/unit/autocomplete.ts
@@ -10,6 +10,7 @@ const CONFIG = { autocomplete: {} };
 const SUGGESTIONS = ['d', 'e', 'f'];
 const NAVIGATIONS = ['g', 'h', 'i'];
 const PRODUCTS = ['j', 'k', 'l', 'm'];
+const PAST_SEARCHES = ['o', 'p', 'q'];
 
 suite('Autocomplete', ({ expect, spy, stub, itShouldProvideAlias }) => {
   let autocomplete: Autocomplete;
@@ -19,6 +20,11 @@ suite('Autocomplete', ({ expect, spy, stub, itShouldProvideAlias }) => {
   beforeEach(() => {
     Autocomplete.prototype.flux = <any>{};
     select = Autocomplete.prototype.select = stub();
+    Autocomplete.prototype.services = <any>{
+      search: {
+        getPastSearchTerms: () => PAST_SEARCHES,
+      },
+    };
     select.withArgs(Selectors.isFetching, 'search').returns(false);
     select.withArgs(Selectors.autocompleteSuggestions).returns(SUGGESTIONS);
     select.withArgs(Selectors.autocompleteCategoryField).returns(CATEGORY);
@@ -46,6 +52,7 @@ suite('Autocomplete', ({ expect, spy, stub, itShouldProvideAlias }) => {
           categoryValues: CATEGORY_VALUES,
           suggestions: SUGGESTIONS,
           navigations: NAVIGATIONS,
+          pastSearches: PAST_SEARCHES,
           products: PRODUCTS,
           isHovered: false,
         });
@@ -112,7 +119,10 @@ suite('Autocomplete', ({ expect, spy, stub, itShouldProvideAlias }) => {
     it('should listen for flux events', () => {
       autocomplete.init();
 
-      expect(on).to.be.calledWith(Events.AUTOCOMPLETE_SUGGESTIONS_UPDATED, autocomplete.updateSuggestions);
+      expect(on)
+        .to.have.callCount(2)
+        .and.calledWith(Events.AUTOCOMPLETE_SUGGESTIONS_UPDATED, autocomplete.updateSuggestions)
+        .and.calledWith(Events.DONE_SEARCH, autocomplete.updatePastSearches);
       expect(subscribe)
         .to.have.callCount(3)
         .and.calledWith('sayt:activate_next', autocomplete.activateNext)
@@ -248,6 +258,20 @@ suite('Autocomplete', ({ expect, spy, stub, itShouldProvideAlias }) => {
       autocomplete.isActive = () => false;
 
       autocomplete.selectActive();
+    });
+  });
+
+  describe('updatePastSearches()', () => {
+    it('should call set() with the past search terms', () => {
+      const pastSearches = ['a', 'b', 'c'];
+      const getPastSearchTerms = spy(() => pastSearches);
+      const set = autocomplete.set = spy();
+      autocomplete.services = <any>{ search: { getPastSearchTerms } };
+
+      autocomplete.updatePastSearches();
+
+      expect(getPastSearchTerms).to.be.called;
+      expect(set).to.be.calledWithExactly({ pastSearches });
     });
   });